+++ src/main/java/me/ajlane/geo/repast/fire/DefaultFireManager.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.List;
 import org.apache.log4j.Logger;
 import cern.jet.random.Poisson;
-import repast.model.agrosuccess.AgroSuccessCodeAliases.Lct;
 import repast.simphony.engine.schedule.ScheduledMethod;
 import repast.simphony.random.RandomHelper;
 import repast.simphony.space.Dimensions;
@@ -18,17 +17,23 @@ public class DefaultFireManager implements FireManager {
   private FireSpreader fireSpreader;
   private Poisson distr;
   private Double fuelMoistureFactor;
+  private final FlammabilityChecker<GridPoint> flammabilityChecker;
 
   /**
-   * @param meanNumFiresPerYear Expected number of fires in the landscape in a year
    * @param fireSpreader Object used to spread fire given an initial ignition
-   * @param fuelMoistureFactor Dimensionless factor parameterising the amount of moisture in the
+   * @param flammabilityChecker Object that indicates whether or not grid cells are flammable (e.g.
+   *        depending on land-cover type)
+   * @param meanNumFiresPerYear Expected number of fires in the landscape in a year
+   * @param vegetationMoistureParam Dimensionless quantity parameterising the amount of moisture in the
    *        fuel at the time of the fire
    */
-  public DefaultFireManager(Double meanNumFiresPerYear, FireSpreader fireSpreader, Double fuelMoistureFactor) {
+  public DefaultFireManager(FireSpreader fireSpreader,
+      FlammabilityChecker<GridPoint> flammabilityChecker,
+      Double meanNumFiresPerYear, Double vegetationMoistureParam) {
     this.fireSpreader = fireSpreader;
+    this.flammabilityChecker = flammabilityChecker;
     this.distr = RandomHelper.createPoisson(meanNumFiresPerYear);
-    this.fuelMoistureFactor = fuelMoistureFactor;
+    this.vegetationMoistureParam = vegetationMoistureParam;
   }
 
   int numFires() {
@@ -48,9 +53,9 @@ public class DefaultFireManager implements FireManager {
       int attemptCounter = 0;
       while (attemptCounter < 1000) {
         GridPoint randomPoint = randomGridPoint(this.fireSpreader.getLct());
-        if (isFlammable(this.fireSpreader.getLct(), randomPoint)) {
+        if (this.flammabilityChecker.isFlammable(randomPoint)) {
           firesStarted.add(randomPoint);
-          this.fireSpreader.spreadFire(randomPoint, this.fuelMoistureFactor);
+          this.fireSpreader.spreadFire(randomPoint, this.vegetationMoistureParam);
           break;
         } else {
           attemptCounter++;
@@ -70,30 +75,6 @@ public class DefaultFireManager implements FireManager {
     int xCoord = RandomHelper.nextIntFromTo(0, (int) dims.getWidth() - 1);
     int yCoord = RandomHelper.nextIntFromTo(0, (int) dims.getHeight() - 1);
     return new GridPoint(xCoord, yCoord);
-  }
-
-  /**
-   * TODO: See about refactoring this code, copied from FireSpreader
-   *
-   * @param lctLayer {@code ValueLayer} encoding land cover types
-   * @param gridPoint Point on grid to query for flammability.
-   * @return {@code true} if the cell at {@code gridPoint} has a flammable land cover type.
-   */
-  private static boolean isFlammable(ValueLayer lctLayer, GridPoint gridPoint) {
-    int lctCode = (int) lctLayer.get(gridPoint.getX(), gridPoint.getY());
-    return isFlammable(lctCode);
-  }
-
-  /**
-   * TODO: See about refactoring this code, copied from FireSpreader
-   *
-   * @param lctCode Code for land cover type
-   * @return {@code true} if {@code lctCode} corresponds to a flammable land cover type.
-   */
-  private static boolean isFlammable(int lctCode) {
-    boolean isWaterQuarry = Lct.WaterQuarry.getCode() == lctCode;
-    boolean isBurnt = Lct.Burnt.getCode() == lctCode;
-    return !(isWaterQuarry || isBurnt); // true if neither water/quarry or burnt
   }
 
 }